+++ src/main/java/com/slozic/dater/exceptions/user/UserBlockedException.java
@@ -1,0 +1,9 @@
+package com.slozic.dater.exceptions.user;
+
+import com.slozic.dater.exceptions.BusinessException;
+
+public class UserBlockedException extends BusinessException {
+    public UserBlockedException(final String message) {
+        super(message);
+    }
+}