+++ src/main/java/com/slozic/dater/repositories/UserReportRepository.java
@@ -1,0 +1,9 @@
+package com.slozic.dater.repositories;
+
+import com.slozic.dater.models.UserReport;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface UserReportRepository extends JpaRepository<UserReport, UUID> {
+}